backend/pyproject.toml
[project]
name = "ssat-backend"
version = "0.1.0"
description = "SSAT Question Generator Backend API"
authors = [
    {name = "rong2ren", email = "98350815+rong2ren@users.noreply.github.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    # Core dependencies (must be explicit for backend deployment)
    "python-dotenv>=1.0.0",
    "loguru>=0.7.3,<0.8.0",
    
    # FastAPI backend dependencies
    "fastapi>=0.116.0",
    "uvicorn>=0.35.0",
    "python-multipart>=0.0.20",
    
    # Data validation and settings
    "pydantic>=2.0.0",
    "pydantic-settings>=2.9.1,<3.0.0",
    "email-validator>=2.0.0",
    "PyJWT>=2.0.0",
    
    # Database and external services
    "supabase>=2.15.1,<3.0.0",
    
    # LLM providers
    "openai>=1.0.0",
    "google-generativeai>=0.3.0",
    
    # ML and embeddings (required for question generation)
    "sentence-transformers>=4.1.0,<5.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.24.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "pytest-asyncio>=1.0.0",
]
